firefly: update mk-openwrt.sh

+++ common/mk-openwrt.sh
@@ -3,14 +3,26 @@ OPENWRT_VERSION=$1
 OPENWRT_DEFCOFNIG=$2
 
 TOP_DIR=$(pwd)
-cd "$TOP_DIR"/openwrt_sdk/"$OPENWRT_VERSION"
+OPENWRT_DIR="$TOP_DIR/openwrt_sdk/$OPENWRT_VERSION"
+OPENWRT_CONFIG_PATH="$OPENWRT_DIR/configs/$OPENWRT_DEFCOFNIG"
+
+if [ ! -d "$OPENWRT_DIR" ];then
+        echo "$OPENWRT_DIR is not exist"
+        exit 1
+fi
+
+if [ ! -e "$OPENWRT_CONFIG_PATH" ];then
+        echo "$OPENWRT_CONFIG_PATH is not exist"
+        exit 1
+fi
+
+cd $OPENWRT_DIR
 
 ./scripts/feeds update -a
 ./scripts/feeds install -a
 
 if [ ! -f .config ]; then
-	cp "$TOP_DIR"/configs/"$OPENWRT_DEFCONFIG" .config
-	make defconfig
+	cp "$OPENWRT_CONFIG_PATH" .config && make defconfig
 else
 	echo "using .config file"
 fi
@@ -20,3 +32,6 @@ find dl -size -1024c -exec ls -l {} \;
 find dl -size -1024c -exec rm -f {} \;
 
 make -j$(nproc) V=s
+RET=$?
+cd -
+exit $RET
